api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
python-multipart==0.0.20
torch==2.7.0
numpy==2.2.5
pillow==11.2.1
opencv-contrib-python==4.11.0.86